fixed an issue with critical attribute checking
